[#2065] Move autoemails bulk email around
This will prevent circular-import in future once typings are added
(next commit).
