+++ amy/consents/util.py
@@ -1,4 +1,5 @@
-from autoemails.actions import NewConsentRequiredAction, send_bulk_email
+from autoemails.actions import NewConsentRequiredAction
+from autoemails.bulk_email import send_bulk_email
 from autoemails.models import Trigger
 from consents.models import Consent, Term
 from workshops.models import Person
